Re-factoring to fully disentangle data and app drops
