+++ daliuge-engine/dlg/apps/dockerapp.py
@@ -40,7 +40,8 @@ import docker
 from docker.models.containers import Container
 
 from dlg import utils, droputils
-from dlg.drop import BarrierAppDROP
+from dlg.named_port_utils import replace_named_ports
+from dlg.apps.app_base import BarrierAppDROP
 from dlg.exceptions import DaliugeException, InvalidDropException
 
 
@@ -241,7 +242,9 @@ class DockerApp(BarrierAppDROP):
         self._cmdLineArgs = self._popArg(kwargs, "command_line_arguments", "")
         self._applicationArgs = self._popArg(kwargs, "applicationArgs", {})
         self._argumentPrefix = self._popArg(kwargs, "argumentPrefix", "--")
-        self._paramValueSeparator = self._popArg(kwargs, "paramValueSeparator", " ")
+        self._paramValueSeparator = self._popArg(
+            kwargs, "paramValueSeparator", " "
+        )
         if not self._image:
             raise InvalidDropException(
                 self, "No docker image specified, cannot create DockerApp"
@@ -258,13 +261,16 @@ class DockerApp(BarrierAppDROP):
 
         self._noBash = False
         if not self._command or self._command.strip()[:2] == "%%":
-            logger.warning("Assume a default command is executed in the container")
-            self._command = self._command.strip().strip()[:2] if self._command else ""
+            logger.warning(
+                "Assume a default command is executed in the container"
+            )
+            self._command = (
+                self._command.strip().strip()[:2] if self._command else ""
+            )
             self._noBash = True
             # This makes sure that we can retain any command defined in the image, but still be
             # able to add any arguments straight after. This requires to use the placeholder string
             # "%%" at the start of the command, else it is interpreted as a normal command.
-
 
         # The user used to run the process in the docker container is now always the user
         # who originally started the DALiuGE process as well. The information is passed through
@@ -338,7 +344,9 @@ class DockerApp(BarrierAppDROP):
             start = time.time()
             c.images.pull(self._image)
             end = time.time()
-            logger.debug("Took %.2f [s] to pull image '%s'", (end - start), self._image)
+            logger.debug(
+                "Took %.2f [s] to pull image '%s'", (end - start), self._image
+            )
         else:
             logger.debug("Image '%s' found, no need to pull it", self._image)
 
@@ -347,12 +355,20 @@ class DockerApp(BarrierAppDROP):
         # If none is provided use the session directory
         inspection = c.api.inspect_image(self._image)
         logger.debug("Docker Image inspection: %r", inspection)
-        self.workdir = inspection.get("ContainerConfig", {}).get("WorkingDir", None)
+        self.workdir = inspection.get("ContainerConfig", {}).get(
+            "WorkingDir", None
+        )
         # self.workdir = None
-        self._sessionId = self._dlg_session.sessionId if self._dlg_session else ""
+        self._sessionId = (
+            self._dlg_session.sessionId if self._dlg_session else ""
+        )
         if not self.workdir:
-            default_workingdir = os.path.join(utils.getDlgWorkDir(), self._sessionId)
-            self.workdir = self._popArg(kwargs, "workingDir", default_workingdir)
+            default_workingdir = os.path.join(
+                utils.getDlgWorkDir(), self._sessionId
+            )
+            self.workdir = self._popArg(
+                kwargs, "workingDir", default_workingdir
+            )
 
         c.api.close()
 
@@ -410,30 +426,41 @@ class DockerApp(BarrierAppDROP):
                 uid: DockerPath(o.path)
                 for uid, o in fsOutputs.items()
             }
-            dataURLInputs = {uid: i for uid, i in iitems if not droputils.has_path(i)}
-            dataURLOutputs = {uid: o for uid, o in oitems if not droputils.has_path(o)}
+            dataURLInputs = {
+                uid: i for uid, i in iitems if not droputils.has_path(i)
+            }
+            dataURLOutputs = {
+                uid: o for uid, o in oitems if not droputils.has_path(o)
+            }
 
             # We bind the inputs and outputs inside the docker under the utils.getDlgDir()
             # directory, maintaining the rest of their original paths.
             # Outputs are bound only up to their dirname (see class doc for details)
             # Volume bindings are setup for FileDROPs and DirectoryContainers only
             binds = [
-                i.path + ":" + dockerInputs[uid].path for uid, i in fsInputs.items()
+                i.path + ":" + dockerInputs[uid].path
+                for uid, i in fsInputs.items()
             ]
             binds += [
-                os.path.dirname(o.path) + ":" + os.path.dirname(dockerOutputs[uid].path)
+                os.path.dirname(o.path)
+                + ":"
+                + os.path.dirname(dockerOutputs[uid].path)
                 for uid, o in fsOutputs.items()
             ]
             logger.debug("Input/output bindings: %r", binds)
             if (
                 len(self._additionalBindings.items()) > 0
             ):  # else we end up with a ':' in the mounts list
-                logger.debug("Additional bindings: %r", self._additionalBindings)
+                logger.debug(
+                    "Additional bindings: %r", self._additionalBindings
+                )
                 binds += [
                     host_path + ":" + container_path
                     for host_path, container_path in self._additionalBindings.items()
                 ]
-            binds = list(set(binds))  # make this a unique list else docker complains
+            binds = list(
+                set(binds)
+            )  # make this a unique list else docker complains
             try:
                 binds.remove(":")
             except:
@@ -448,7 +475,9 @@ class DockerApp(BarrierAppDROP):
                     else:
                         host_port, container_port = mapping.split(":")
                     if host_port not in portMappings:
-                        logger.debug(f"mapping port {host_port} -> {container_port}")
+                        logger.debug(
+                            f"mapping port {host_port} -> {container_port}"
+                        )
                         portMappings[host_port] = int(container_port)
                     else:
                         raise Exception(
@@ -464,7 +493,9 @@ class DockerApp(BarrierAppDROP):
             if self._user is not None:
                 env.update({"USER": self._user, "DLG_ROOT": utils.getDlgDir()})
             if self._env is not None:
-                logger.debug(f"Found environment variable setting: {self._env}")
+                logger.debug(
+                    f"Found environment variable setting: {self._env}"
+                )
                 if (
                     self._env.lower() == "all"
                 ):  # pass on all environment variables from host
@@ -477,7 +508,9 @@ class DockerApp(BarrierAppDROP):
                             "Ignoring provided environment variables: Format wrong? Check documentation"
                         )
                         addEnv = {}
-                    if isinstance(addEnv, dict):  # if it is a dict populate directly
+                    if isinstance(
+                        addEnv, dict
+                    ):  # if it is a dict populate directly
                         # but replace placeholders first
                         for key in addEnv:
                             value = droputils.replace_path_placeholders(
@@ -501,17 +534,28 @@ class DockerApp(BarrierAppDROP):
 
             # deal with named ports
             appArgs = self._applicationArgs
-            inport_names = self.parameters['inputs'] \
-                if "inputs" in self.parameters else []
-            outport_names = self.parameters['outputs'] \
-                if "outputs" in self.parameters else []
-            keyargs, pargs = droputils.replace_named_ports(iitems, oitems, 
-                inport_names, outport_names, appArgs, 
-                argumentPrefix=self._argumentPrefix, 
-                separator=self._paramValueSeparator)
+            inport_names = (
+                self.parameters["inputs"]
+                if "inputs" in self.parameters
+                else []
+            )
+            outport_names = (
+                self.parameters["outputs"]
+                if "outputs" in self.parameters
+                else []
+            )
+            keyargs, pargs = replace_named_ports(
+                iitems,
+                oitems,
+                inport_names,
+                outport_names,
+                appArgs,
+                argumentPrefix=self._argumentPrefix,
+                separator=self._paramValueSeparator,
+            )
 
             argumentString = f"{' '.join(keyargs + pargs)}"
-            
+
             # complete command including all additional parameters and optional redirects
             cmd = f"{self._command} {argumentString} {self._cmdLineArgs} "
             if cmd:
@@ -536,7 +580,9 @@ class DockerApp(BarrierAppDROP):
                 cmd = '/bin/bash -c "%s"' % (
                     utils.escapeQuotes(cmd, singleQuotes=False)
                 )
-                logger.debug("Command after user creation and wrapping is: %s", cmd)
+                logger.debug(
+                    "Command after user creation and wrapping is: %s", cmd
+                )
             else:
                 logger.debug(
                     "executing container with default cmd and wrapped arguments"
@@ -544,7 +590,9 @@ class DockerApp(BarrierAppDROP):
                 cmd = f"{utils.escapeQuotes(cmd, singleQuotes=False)}"
 
             c = DockerApp._get_client()
-            logger.debug(f"Final user for container: {self._user}:{self._userid}")
+            logger.debug(
+                f"Final user for container: {self._user}:{self._userid}"
+            )
 
             # Create container
             self._container = c.containers.create(  # type: ignore
@@ -595,8 +643,12 @@ class DockerApp(BarrierAppDROP):
             end = time.time()
 
             # Capture output
-            stdout = self.container.logs(stream=False, stdout=True, stderr=False)
-            stderr = self.container.logs(stream=False, stdout=False, stderr=True)
+            stdout = self.container.logs(
+                stream=False, stdout=True, stderr=False
+            )
+            stderr = self.container.logs(
+                stream=False, stdout=False, stderr=True
+            )
             if isinstance(stdout, bytes):
                 stdout = stdout.decode()
                 stderr = stderr.decode()
