+++ daliuge-engine/dlg/apps/simple.py
@@ -33,7 +33,8 @@ import ast
 import numpy as np
 
 from dlg import droputils, utils
-from dlg.drop import BarrierAppDROP, ContainerDROP
+from dlg.apps.app_base import BarrierAppDROP
+from dlg.data.drops.container import ContainerDROP
 from dlg.apps.branch import BranchAppDrop
 from dlg.meta import (
     dlg_float_param,
