+++ daliuge-engine/dlg/apps/socket_listener.py
@@ -29,7 +29,7 @@ import logging
 import socket
 
 from ..ddap_protocol import DROPRel, DROPLinkType
-from ..drop import BarrierAppDROP
+from ..apps.app_base import BarrierAppDROP
 from ..exceptions import InvalidRelationshipException
 from ..meta import (
     dlg_string_param,
@@ -102,7 +102,9 @@ class SocketListenerApp(BarrierAppDROP):
         # At least one output should have been added
         outs = self.outputs
         if len(outs) < 1:
-            raise Exception("At least one output should have been added to %r" % self)
+            raise Exception(
+                "At least one output should have been added to %r" % self
+            )
 
         # Don't really listen for data if running dry
         if self._dryRun:
@@ -112,14 +114,18 @@ class SocketListenerApp(BarrierAppDROP):
         serverSocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         with contextlib.closing(serverSocket):
             if self.reuseAddr:
-                serverSocket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+                serverSocket.setsockopt(
+                    socket.SOL_SOCKET, socket.SO_REUSEADDR, 1
+                )
             serverSocket.bind((self.host, self.port))
             serverSocket.listen(1)
             logger.debug(
                 "Listening for a TCP connection on %s:%d", self.host, self.port
             )
             clientSocket, address = serverSocket.accept()
-            logger.info("Accepted connection from %s:%d", address[0], address[1])
+            logger.info(
+                "Accepted connection from %s:%d", address[0], address[1]
+            )
 
         # Simply write the data we receive into our outputs
         n = 0
@@ -142,6 +148,8 @@ class SocketListenerApp(BarrierAppDROP):
 
     def addStreamingInput(self, streamingInputDrop, back=True):
         raise InvalidRelationshipException(
-            DROPRel(streamingInputDrop.uid, DROPLinkType.STREAMING_INPUT, self.uid),
+            DROPRel(
+                streamingInputDrop.uid, DROPLinkType.STREAMING_INPUT, self.uid
+            ),
             "SocketListenerApp should have no inputs",
         )
