+++ daliuge-engine/dlg/data/drops/directorycontainer.py
@@ -25,7 +25,8 @@ import shutil
 
 from dlg.data.drops.file import FileDROP
 from dlg.ddap_protocol import DROPRel, DROPLinkType
-from dlg.drop import PathBasedDrop, ContainerDROP
+from dlg.data.drops.data_base import PathBasedDrop
+from dlg.data.drops.container import ContainerDROP
 from dlg.exceptions import InvalidDropException, InvalidRelationshipException
 from dlg.meta import dlg_bool_param
 
@@ -67,10 +68,14 @@ class DirectoryContainer(PathBasedDrop, ContainerDROP):
 
         directory = kwargs["dirname"]
 
-        logger.debug("Checking existence of %s %s", directory, self.check_exists)
+        logger.debug(
+            "Checking existence of %s %s", directory, self.check_exists
+        )
         if "check_exists" in kwargs and kwargs["check_exists"] is True:
             if not os.path.isdir(directory):
-                raise InvalidDropException(self, "%s is not a directory" % (directory))
+                raise InvalidDropException(
+                    self, "%s is not a directory" % (directory)
+                )
 
         self._path = self.get_dir(directory)
 
@@ -84,7 +89,9 @@ class DirectoryContainer(PathBasedDrop, ContainerDROP):
                 )
             ContainerDROP.addChild(self, child)
         else:
-            raise TypeError("Child DROP is not of type FileDROP or DirectoryContainer")
+            raise TypeError(
+                "Child DROP is not of type FileDROP or DirectoryContainer"
+            )
 
     def delete(self):
         shutil.rmtree(self._path)
