+++ daliuge-engine/dlg/data/drops/s3_drop.py
@@ -35,10 +35,11 @@ try:
 except ImportError:
     logger.warning("BOTO bindings are not available")
 
-from ...drop import DataDROP
+from ...data.drops.data_base import DataDROP
 from dlg.data.io import ErrorIO, OpenMode, DataIO
 from ...meta import dlg_string_param, dlg_list_param
-from dlg.droputils import identify_named_ports, check_ports_dict
+
+from dlg.named_port_utils import identify_named_ports, check_ports_dict
 
 
 ##
@@ -53,7 +54,7 @@ from dlg.droputils import identify_named_ports, check_ports_dict
 # @param object_name Object Name//String/ComponentParameter/readwrite//False/False/The S3 object key
 # @param profile_name Profile Name//String/ComponentParameter/readwrite//False/False/The S3 profile name
 # @param endpoint_url Endpoint URL//String/ComponentParameter/readwrite//False/False/The URL exposing the S3 REST API
-# @param streaming Streaming/False/Boolean/ComponentParameter/readwrite//False/False/Specifies whether this data component streams input and output data 
+# @param streaming Streaming/False/Boolean/ComponentParameter/readwrite//False/False/Specifies whether this data component streams input and output data
 # @param persist Persist/False/Boolean/ComponentParameter/readwrite//False/False/Specifies whether this data component contains data that should not be deleted after execution
 # @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
 # @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
@@ -82,7 +83,7 @@ class S3DROP(DataDROP):
             "aws_access_key_id": self.aws_access_key_id,
             "aws_secret_access_key": self.aws_secret_access_key,
             "profile_name": self.profile_name,
-            "endpoint_url": self.endpoint_url
+            "endpoint_url": self.endpoint_url,
         }
         self.Key = self.uid if not self.Key else self.Key
         return super().initialize(**kwargs)
@@ -107,49 +108,58 @@ class S3DROP(DataDROP):
 
     def getIO(self) -> DataIO:
         """
-        Return 
+        Return
         :return:
         """
         logger.debug("S3DROP producers: %s", self._producers)
         if check_ports_dict(self._producers):
             self.mapped_inputs = identify_named_ports(
-                self._producers, 
-                {},
-                self.keyargs,
-                mode="inputs"
-                )
+                self._producers, {}, self.keyargs, mode="inputs"
+            )
         logger.debug("Parameters found: {}", self.parameters)
-        return S3IO(self.aws_access_key_id,
-                    self.aws_secret_access_key,
-                    self.profile_name,
-                    self.Bucket,
-                    self.Key,
-                    self.endpoint_url,
-                    self._expectedSize,
-                    )
+        return S3IO(
+            self.aws_access_key_id,
+            self.aws_secret_access_key,
+            self.profile_name,
+            self.Bucket,
+            self.Key,
+            self.endpoint_url,
+            self._expectedSize,
+        )
 
 
 class S3IO(DataIO):
     """
     IO class for the S3 Drop
     """
+
     _desc = None
 
-    def __init__(self,
-                 aws_access_key_id=None,
-                 aws_secret_access_key=None,
-                 profile_name=None,
-                 Bucket=None,
-                 Key=None,
-                 endpoint_url=None,
-                 expectedSize=-1,
-                 **kwargs):
+    def __init__(
+        self,
+        aws_access_key_id=None,
+        aws_secret_access_key=None,
+        profile_name=None,
+        Bucket=None,
+        Key=None,
+        endpoint_url=None,
+        expectedSize=-1,
+        **kwargs,
+    ):
 
         super().__init__(**kwargs)
 
-        logger.debug(("key_id: %s; key: %s; profile: %s; bucket: %s; object_id: %s; %s",
-                      aws_access_key_id, aws_secret_access_key, profile_name, Bucket, Key,
-                      endpoint_url))
+        logger.debug(
+            (
+                "key_id: %s; key: %s; profile: %s; bucket: %s; object_id: %s; %s",
+                aws_access_key_id,
+                aws_secret_access_key,
+                profile_name,
+                Bucket,
+                Key,
+                endpoint_url,
+            )
+        )
         self._s3 = None
         self._s3_access_key_id = aws_access_key_id
         self._s3_secret_access_key = aws_secret_access_key
@@ -161,7 +171,7 @@ class S3IO(DataIO):
         self._s3 = self._get_s3_connection()
         self.url = f"{endpoint_url}/{Bucket}/{Key}"
         self._expectedSize = expectedSize
-        self._buffer = b''
+        self._buffer = b""
         if self._mode == 1:
             try:
                 self._s3Stream = self._open()
@@ -173,14 +183,12 @@ class S3IO(DataIO):
     def _get_s3_connection(self):
         s3 = None
         if self._s3 is None:
-            if (
-                    self._profile_name is not None or
-                    (self._s3_access_key_id is not None
-                     and self._s3_secret_access_key is not None)
+            if self._profile_name is not None or (
+                self._s3_access_key_id is not None
+                and self._s3_secret_access_key is not None
             ):
                 logger.debug("Opening boto3 session")
-                session = boto3.Session(
-                    profile_name=self._profile_name)
+                session = boto3.Session(profile_name=self._profile_name)
                 s3 = session.client(
                     service_name="s3",
                     endpoint_url=self._s3_endpoint_url,
@@ -216,8 +224,8 @@ class S3IO(DataIO):
             return self._s3
         else:
             s3Object = self._s3.get_object(Bucket=self._bucket, Key=self._key)
-            self._desc = s3Object['Body']
-        return s3Object['Body']
+            self._desc = s3Object["Body"]
+        return s3Object["Body"]
 
     @overrides
     def _read(self, count=-1, **kwargs):
@@ -229,7 +237,7 @@ class S3IO(DataIO):
         else:
             return self._desc.read()
 
-    def _writeBuffer2S3(self, write_buffer=b''):
+    def _writeBuffer2S3(self, write_buffer=b""):
         try:
             with BytesIO(write_buffer) as f:
                 self._s3.upload_part(
@@ -237,11 +245,14 @@ class S3IO(DataIO):
                     Bucket=self._bucket,
                     Key=self._key,
                     UploadId=self._uploadId,
-                    PartNumber=self._partNo)
-            logger.debug("Wrote %d bytes part %d to S3: %s",
-                         len(write_buffer),
-                         self._partNo,
-                         self.url)
+                    PartNumber=self._partNo,
+                )
+            logger.debug(
+                "Wrote %d bytes part %d to S3: %s",
+                len(write_buffer),
+                self._partNo,
+                self.url,
+            )
             self._partNo += 1
             self._written += len(write_buffer)
         except botocore.exceptions.ClientError as e:
@@ -250,10 +261,9 @@ class S3IO(DataIO):
 
     @overrides
     def _write(self, data, **kwargs) -> int:
-        """
-        """
+        """ """
         self._buffer += data
-        PART_SIZE = 5 * 1024 ** 2
+        PART_SIZE = 5 * 1024**2
         logger.debug("Length of S3 buffer: %d", len(self._buffer))
         if len(self._buffer) >= PART_SIZE:
             self._writeBuffer2S3(self._buffer[:PART_SIZE])
@@ -262,40 +272,44 @@ class S3IO(DataIO):
         # to keep the client happy
 
     def _get_object_head(self) -> dict:
-        return self._s3.head_object(
-            Bucket=self._bucket,
-            Key=self._key)
+        return self._s3.head_object(Bucket=self._bucket, Key=self._key)
 
     @overrides
     def _size(self, **kwargs) -> int:
         if self.exists():
             object_head = self._get_object_head()
-            logger.debug(("Size of object:%s", object_head['ContentLength']))
-            return object_head['ContentLength']
+            logger.debug(("Size of object:%s", object_head["ContentLength"]))
+            return object_head["ContentLength"]
         return -1
 
     @overrides
     def _close(self, **kwargs):
         if self._mode == OpenMode.OPEN_WRITE:
-            if len(self._buffer) > 0:  # write, if there is still something in the buffer
+            if (
+                len(self._buffer) > 0
+            ):  # write, if there is still something in the buffer
                 self._writeBuffer2S3(self._buffer)
             # complete multipart upload and cleanup
             res = self._s3.list_parts(
-                Bucket=self._bucket,
-                Key=self._key,
-                UploadId=self._uploadId)
-            parts = [{'ETag': p['ETag'],
-                      'PartNumber': p['PartNumber']} for p in res['Parts']]
+                Bucket=self._bucket, Key=self._key, UploadId=self._uploadId
+            )
+            parts = [
+                {"ETag": p["ETag"], "PartNumber": p["PartNumber"]}
+                for p in res["Parts"]
+            ]
             # TODO: Check checksum!
             res = self._s3.complete_multipart_upload(
                 Bucket=self._bucket,
                 Key=self._key,
                 UploadId=self._uploadId,
-                MultipartUpload={'Parts': parts},
+                MultipartUpload={"Parts": parts},
             )
             del self._buffer
-            logger.info("Wrote a total of %.1f MB to %s",
-                        self._written / (1024 ** 2), self.url)
+            logger.info(
+                "Wrote a total of %.1f MB to %s",
+                self._written / (1024**2),
+                self.url,
+            )
 
         self._desc.close()
         del self._s3
@@ -323,7 +337,11 @@ class S3IO(DataIO):
                 logger.info("Access to bucket %s is forbidden", self._bucket)
                 return False, False
             elif error_code > 300:
-                logger.info("Error code %s when accessing bucket %s", error_code, self._bucket)
+                logger.info(
+                    "Error code %s when accessing bucket %s",
+                    error_code,
+                    self._bucket,
+                )
         try:
             logger.info("Checking existence of object: %s", self._key)
             s3.head_object(Bucket=self._bucket, Key=self._key)
