+++ daliuge-engine/dlg/drop.py
@@ -22,11 +22,7 @@
 """
 Module containing the core DROP classes.
 """
-from sqlite3 import OperationalError
-from abc import ABCMeta, abstractmethod, abstractproperty
 import ast
-from collections import OrderedDict
-import heapq
 import inspect
 import logging
 import math
@@ -36,7 +32,7 @@ import threading
 import time
 import re
 import sys
-from typing import List, Union
+from abc import ABCMeta
 
 from dlg.common.reproducibility.constants import (
     ReproducibilityFlags,
@@ -50,7 +46,6 @@ from merklelib import MerkleTree
 from .ddap_protocol import (
     ExecutionMode,
     ChecksumTypes,
-    AppDROPStates,
     DROPLinkType,
     DROPPhases,
     DROPStates,
@@ -58,12 +53,6 @@ from .ddap_protocol import (
 )
 from dlg.event import EventFirer, EventHandler
 from dlg.exceptions import InvalidDropException, InvalidRelationshipException
-from dlg.data.io import (
-    DataIO,
-    OpenMode,
-    ErrorIO,
-    NullIO,
-)
 
 DEFAULT_INTERNAL_PARAMETERS = {
     "storage",
@@ -85,7 +74,6 @@ from dlg.utils import (
     object_tracking,
     getDlgVariable,
 )
-from dlg.process import DlgProcess
 from dlg.meta import (
     dlg_float_param,
     dlg_int_param,
@@ -1189,909 +1177,6 @@ class AbstractDROP(EventFirer, EventHandler):
     @property
     def parameters(self):
         return self._parameters
-
-
-class PathBasedDrop(object):
-    """
-    Base class for data drops that handle paths (i.e., file and directory drops)
-    """
-
-    _path: str = None
-
-    def get_dir(self, dirname):
-
-        if isabs(dirname):
-            return dirname
-
-        # dirname will be based on the current working directory
-        # If we have a session, it goes into the path as well
-        # (most times we should have a session BTW, we should expect *not* to
-        # have one only during testing)
-        parts = []
-        if self._dlg_session:
-            parts.append(".")
-        else:
-            parts.append("/tmp/daliuge_tfiles")
-        if dirname:
-            parts.append(dirname)
-
-        the_dir = os.path.abspath(os.path.normpath(os.path.join(*parts)))
-        logger.debug("Path used for drop: %s", the_dir)
-        createDirIfMissing(the_dir)
-        return the_dir
-
-    @property
-    def path(self) -> str:
-        return self._path
-
-
-##
-# @brief Data
-# @details A generic Data drop, whose functionality can be provided by an arbitrary class, as specified in the 'dataclass' component parameter
-# @par EAGLE_START
-# @param category Data
-# @param tag template
-# @param dataclass Data Class/my.awesome.data.Component/String/ComponentParameter/readonly//False/False/The python class that implements this data component
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param streaming Streaming/False/Boolean/ComponentParameter/readwrite//False/False/Specifies whether this data component streams input and output data
-# @param persist Persist/False/Boolean/ComponentParameter/readwrite//False/False/Specifies whether this data component contains data that should not be deleted after execution
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class DataDROP(AbstractDROP):
-    """
-    A DataDROP is a DROP that stores data for writing with
-    an AppDROP, or reading with one or more AppDROPs.
-
-    DataDROPs have two different modes: "normal" and "streaming".
-    Normal DataDROPs will wait until the COMPLETED state before being
-    available as input to an AppDROP, while streaming AppDROPs may be
-    read simutaneously with writing by chunking drop bytes together.
-
-    This class contains two methods that need to be overrwritten:
-    `getIO`, invoked by AppDROPs when reading or writing to a drop,
-    and `dataURL`, a getter for a data URI uncluding protocol and address
-    parsed by function `IOForURL`.
-    """
-
-    def incrRefCount(self):
-        """
-        Increments the reference count of this DROP by one atomically.
-        """
-        with self._refLock:
-            self._refCount += 1
-
-    def decrRefCount(self):
-        """
-        Decrements the reference count of this DROP by one atomically.
-        """
-        with self._refLock:
-            self._refCount -= 1
-
-    @track_current_drop
-    def open(self, **kwargs):
-        """
-        Opens the DROP for reading, and returns a "DROP descriptor"
-        that must be used when invoking the read() and close() methods.
-        DROPs maintain a internal reference count based on the number
-        of times they are opened for reading; because of that after a successful
-        call to this method the corresponding close() method must eventually be
-        invoked. Failing to do so will result in DROPs not expiring and
-        getting deleted.
-        """
-        if self.status != DROPStates.COMPLETED:
-            raise Exception(
-                "%r is in state %s (!=COMPLETED), cannot be opened for reading"
-                % (self, self.status)
-            )
-
-        io = self.getIO()
-        logger.debug("Opening drop %s", self.oid)
-        io.open(OpenMode.OPEN_READ, **kwargs)
-
-        # Save the IO object in the dictionary and return its descriptor instead
-        while True:
-            descriptor = random.SystemRandom().randint(-(2**31), 2**31 - 1)
-            if descriptor not in self._rios:
-                break
-        self._rios[descriptor] = io
-
-        # This occurs only after a successful opening
-        self.incrRefCount()
-        self._fire("open")
-
-        return descriptor
-
-    @track_current_drop
-    def close(self, descriptor, **kwargs):
-        """
-        Closes the given DROP descriptor, decreasing the DROP's
-        internal reference count and releasing the underlying resources
-        associated to the descriptor.
-        """
-        self._checkStateAndDescriptor(descriptor)
-
-        # Decrement counter and then actually close
-        self.decrRefCount()
-        io = self._rios.pop(descriptor)
-        io.close(**kwargs)
-
-    def _closeWriters(self):
-        """
-        Close our writing IO instance.
-        If written externally, self._wio will have remained None
-        """
-        if self._wio:
-            try:
-                self._wio.close()
-            except:
-                pass  # this will make sure that a previous issue does not cause the graph to hang!
-                # raise Exception("Problem closing file!")
-            self._wio = None
-
-    def read(self, descriptor, count=65536, **kwargs):
-        """
-        Reads `count` bytes from the given DROP `descriptor`.
-        """
-        self._checkStateAndDescriptor(descriptor)
-        io = self._rios[descriptor]
-        return io.read(count, **kwargs)
-
-    def _checkStateAndDescriptor(self, descriptor):
-        if self.status != DROPStates.COMPLETED:
-            raise Exception(
-                "%r is in state %s (!=COMPLETED), cannot be read"
-                % (self, self.status)
-            )
-        if descriptor is None:
-            raise ValueError("Illegal empty descriptor given")
-        if descriptor not in self._rios:
-            raise Exception(
-                "Illegal descriptor %d given, remember to open() first"
-                % (descriptor)
-            )
-
-    def isBeingRead(self):
-        """
-        Returns `True` if the DROP is currently being read; `False`
-        otherwise
-        """
-        with self._refLock:
-            return self._refCount > 0
-
-    @track_current_drop
-    def write(self, data: Union[bytes, memoryview], **kwargs):
-        """
-        Writes the given `data` into this DROP. This method is only meant
-        to be called while the DROP is in INITIALIZED or WRITING state;
-        once the DROP is COMPLETE or beyond only reading is allowed.
-        The underlying storage mechanism is responsible for implementing the
-        final writing logic via the `self.writeMeta()` method.
-        """
-
-        if self.status not in [DROPStates.INITIALIZED, DROPStates.WRITING]:
-            raise Exception("No more writing expected")
-
-        if not isinstance(data, (bytes, memoryview)):
-            raise Exception(
-                "Data type not of binary type: %s", type(data).__name__
-            )
-
-        # We lazily initialize our writing IO instance because the data of this
-        # DROP might not be written through this DROP
-        if not self._wio:
-            self._wio = self.getIO()
-            try:
-                self._wio.open(OpenMode.OPEN_WRITE)
-            except:
-                self.status = DROPStates.ERROR
-                raise Exception("Problem opening drop for write!")
-        nbytes = self._wio.write(data)
-        nbytes = 0 if nbytes is None else nbytes
-
-        dataLen = len(data)
-        if nbytes != dataLen:
-            # TODO: Maybe this should be an actual error?
-            logger.warning(
-                "Not all data was correctly written by %s (%d/%d bytes written)",
-                self,
-                nbytes,
-                dataLen,
-            )
-
-        # see __init__ for the initialization to None
-        if self._size is None:
-            self._size = 0
-        self._size += nbytes
-
-        # Trigger our streaming consumers
-        if self._streamingConsumers:
-            for streamingConsumer in self._streamingConsumers:
-                streamingConsumer.dataWritten(self.uid, data)
-
-        # Update our internal checksum
-        if not checksum_disabled:
-            self._updateChecksum(data)
-
-        # If we know how much data we'll receive, keep track of it and
-        # automatically switch to COMPLETED
-        if self._expectedSize > 0:
-            remaining = self._expectedSize - self._size
-            if remaining > 0:
-                self.status = DROPStates.WRITING
-            else:
-                if remaining < 0:
-                    logger.warning(
-                        "Received and wrote more bytes than expected: %d",
-                        -remaining,
-                    )
-                logger.debug(
-                    "Automatically moving %r to COMPLETED, all expected data arrived",
-                    self,
-                )
-                self.setCompleted()
-        else:
-            self.status = DROPStates.WRITING
-
-        return nbytes
-
-    def _updateChecksum(self, chunk):
-        # see __init__ for the initialization to None
-        if self._checksum is None:
-            self._checksum = 0
-            self._checksumType = _checksumType
-        self._checksum = crc32c(chunk, self._checksum)
-
-    @property
-    def checksum(self):
-        """
-        The checksum value for the data represented by this DROP. Its
-        value is automatically calculated if the data was actually written
-        through this DROP (using the `self.write()` method directly or
-        indirectly). In the case that the data has been externally written, the
-        checksum can be set externally after the DROP has been moved to
-        COMPLETED or beyond.
-
-        :see: `self.checksumType`
-        """
-        if self.status == DROPStates.COMPLETED and self._checksum is None:
-            # Generate on the fly
-            io = self.getIO()
-            io.open(OpenMode.OPEN_READ)
-            data = io.read(65536)
-            while data is not None and len(data) > 0:
-                self._updateChecksum(data)
-                data = io.read(65536)
-            io.close()
-        return self._checksum
-
-    @checksum.setter
-    def checksum(self, value):
-        if self._checksum is not None:
-            raise Exception(
-                "The checksum for DROP %s is already calculated, cannot overwrite with new value"
-                % (self)
-            )
-        if self.status in [DROPStates.INITIALIZED, DROPStates.WRITING]:
-            raise Exception(
-                "DROP %s is still not fully written, cannot manually set a checksum yet"
-                % (self)
-            )
-        self._checksum = value
-
-    @property
-    def checksumType(self):
-        """
-        The algorithm used to compute this DROP's data checksum. Its value
-        if automatically set if the data was actually written through this
-        DROP (using the `self.write()` method directly or indirectly). In
-        the case that the data has been externally written, the checksum type
-        can be set externally after the DROP has been moved to COMPLETED
-        or beyond.
-
-        :see: `self.checksum`
-        """
-        return self._checksumType
-
-    @checksumType.setter
-    def checksumType(self, value):
-        if self._checksumType is not None:
-            raise Exception(
-                "The checksum type for DROP %s is already set, cannot overwrite with new value"
-                % (self)
-            )
-        if self.status in [DROPStates.INITIALIZED, DROPStates.WRITING]:
-            raise Exception(
-                "DROP %s is still not fully written, cannot manually set a checksum type yet"
-                % (self)
-            )
-        self._checksumType = value
-
-    @abstractmethod
-    def getIO(self) -> DataIO:
-        """
-        Returns an instance of one of the `dlg.io.DataIO` instances that
-        handles the data contents of this DROP.
-        """
-
-    def delete(self):
-        """
-        Deletes the data represented by this DROP.
-        """
-        self.getIO().delete()
-
-    def exists(self):
-        """
-        Returns `True` if the data represented by this DROP exists indeed
-        in the underlying storage mechanism
-        """
-        return self.getIO().exists()
-
-    @abstractproperty
-    def dataURL(self) -> str:
-        """
-        A URL that points to the data referenced by this DROP. Different
-        DROP implementations will use different URI schemes.
-        """
-
-
-##
-# @brief NULL
-# @details A Drop not storing any data (useful for just passing on events)
-# @par EAGLE_START
-# @param category Memory
-# @param tag daliuge
-# @param data_volume Data volume/0/Float/ComponentParameter/readonly//False/False/This never stores any data
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class NullDROP(DataDROP):
-    """
-    A DROP that doesn't store any data.
-    """
-
-    def getIO(self):
-        return NullIO()
-
-    @property
-    def dataURL(self) -> str:
-        return "null://"
-
-
-class EndDROP(NullDROP):
-    """
-    A DROP that ends the session when reached
-    """
-
-
-class ContainerDROP(DataDROP):
-    """
-    A DROP that doesn't directly point to some piece of data, but instead
-    holds references to other DROPs (its children), and from them its own
-    internal state is deduced.
-
-    Because of its nature, ContainerDROPs cannot be written to directly,
-    and likewise they cannot be read from directly. One instead has to pay
-    attention to its "children" DROPs if I/O must be performed.
-    """
-
-    def initialize(self, **kwargs):
-        super(DataDROP, self).initialize(**kwargs)
-        self._children = []
-
-    # ===========================================================================
-    # No data-related operations should actually be called in Container DROPs
-    # ===========================================================================
-    def getIO(self):
-        return ErrorIO()
-
-    @property
-    def dataURL(self):
-        raise OperationalError()
-
-    def addChild(self, child):
-
-        # Avoid circular dependencies between Containers
-        if child == self.parent:
-            raise InvalidRelationshipException(
-                DROPRel(child, DROPLinkType.CHILD, self),
-                "Circular dependency found",
-            )
-
-        logger.debug("Adding new child for %r: %r", self, child)
-
-        self._children.append(child)
-        child.parent = self
-
-    def delete(self):
-        # TODO: this needs more thinking. Probably a separate method to perform
-        #       this recursive deletion will be needed, while this delete method
-        #       will go hand-to-hand with the rest of the I/O methods above,
-        #       which are currently raise a NotImplementedError
-        if self._children:
-            for c in [c for c in self._children if c.exists()]:
-                c.delete()
-
-    @property
-    def expirationDate(self):
-        if self._children:
-            return heapq.nlargest(
-                1, [c.expirationDate for c in self._children]
-            )[0]
-        return self._expirationDate
-
-    @property
-    def children(self):
-        return self._children[:]
-
-    def exists(self):
-        if self._children:
-            # TODO: Or should it be all()? Depends on what the exact contract of
-            #       "exists" is
-            return any([c.exists() for c in self._children])
-        return True
-
-
-# ===============================================================================
-# AppDROP classes follow
-# ===============================================================================
-
-
-class AppDROP(ContainerDROP):
-    """
-    An AppDROP is a DROP representing an application that reads data
-    from one or more DataDROPs (its inputs), and writes data onto one or more
-    DataDROPs (its outputs).
-
-    AppDROPs accept two different kind of inputs: "normal" and "streaming"
-    inputs. Normal inputs are DataDROPs that must be on the COMPLETED state
-    (and therefore their data must be fully written) before this application is
-    run, while streaming inputs are DataDROPs that feed chunks of data into
-    this application as the data gets written into them.
-
-    This class contains two methods that need to be overwritten by
-    subclasses: `dropCompleted`, invoked when input DataDROPs move to
-    COMPLETED, and `dataWritten`, invoked with the data coming from streaming
-    inputs.
-
-    How and when applications are executed is completely up to the app component
-    developer, and is not enforced by this base class. Some applications might need
-    to be run at `initialize` time, while other might start during the first invocation
-    of `dataWritten`. A common scenario anyway is to start an application only
-    after all its inputs have moved to COMPLETED (implying that none of them is
-    an streaming input); for these cases see the `BarrierAppDROP`.
-    """
-
-    def initialize(self, **kwargs):
-
-        super(AppDROP, self).initialize(**kwargs)
-
-        # Inputs and Outputs are the DROPs that get read from and written
-        # to by this AppDROP, respectively. An input DROP will see
-        # this AppDROP as one of its consumers, while an output DROP
-        # will see this AppDROP as one of its producers.
-        #
-        # Input and output objects are later referenced by their *index*
-        # (relative to the order in which they were added to this object)
-        # Therefore we use an ordered dict to keep the insertion order.
-        self._inputs = OrderedDict()
-        self._outputs = OrderedDict()
-
-        # Same as above, only that these correspond to the 'streaming' version
-        # of the consumers
-        self._streamingInputs = OrderedDict()
-
-        # An AppDROP has a second, separate state machine indicating its
-        # execution status.
-        self._execStatus = AppDROPStates.NOT_RUN
-
-    @track_current_drop
-    def addInput(self, inputDrop, back=True):
-        uid = inputDrop.uid
-        if uid not in self._inputs:
-            self._inputs[uid] = inputDrop
-            if back:
-                inputDrop.addConsumer(self, False)
-
-    @property
-    def inputs(self) -> List[DataDROP]:
-        """
-        The list of inputs set into this AppDROP
-        """
-        return list(self._inputs.values())
-
-    @track_current_drop
-    def addOutput(self, outputDrop: DataDROP, back=True):
-        if outputDrop is self:
-            raise InvalidRelationshipException(
-                DROPRel(outputDrop, DROPLinkType.OUTPUT, self),
-                "Cannot add an AppConsumer as its own output",
-            )
-        uid = outputDrop.uid
-        if uid not in self._outputs:
-            self._outputs[uid] = outputDrop
-
-            if back:
-                outputDrop.addProducer(self, False)
-
-            # Subscribe the output DROP to events sent by this AppDROP when it
-            # finishes its execution.
-            self.subscribe(outputDrop, "producerFinished")
-
-    @property
-    def outputs(self) -> List[DataDROP]:
-        """
-        The list of outputs set into this AppDROP
-        """
-        return list(self._outputs.values())
-
-    def addStreamingInput(self, streamingInputDrop, back=True):
-        if streamingInputDrop not in self._streamingInputs.values():
-            uid = streamingInputDrop.uid
-            self._streamingInputs[uid] = streamingInputDrop
-            if back:
-                streamingInputDrop.addStreamingConsumer(self, False)
-
-    @property
-    def streamingInputs(self) -> List[DataDROP]:
-        """
-        The list of streaming inputs set into this AppDROP
-        """
-        return list(self._streamingInputs.values())
-
-    def _generateNamedInputs(self):
-        """
-        Generates a named mapping of input data drops. Can only be called during run().
-        """
-        named_inputs: OrderedDict[str, DataDROP] = OrderedDict()
-        if "inputs" in self.parameters and isinstance(
-            self.parameters["inputs"][0], dict
-        ):
-            for i in range(len(self._inputs)):
-                key = list(self.parameters["inputs"][i].values())[0]
-                value = self._inputs[
-                    list(self.parameters["inputs"][i].keys())[0]
-                ]
-                named_inputs[key] = value
-        return named_inputs
-
-    def _generateNamedOutputs(self):
-        """
-        Generates a named mapping of output data drops. Can only be called during run().
-        """
-        named_outputs: OrderedDict[str, DataDROP] = OrderedDict()
-        if "outputs" in self.parameters and isinstance(
-            self.parameters["outputs"][0], dict
-        ):
-            for i in range(len(self._outputs)):
-                key = list(self.parameters["outputs"][i].values())[0]
-                value = self._outputs[
-                    list(self.parameters["outputs"][i].keys())[0]
-                ]
-                named_outputs[key] = value
-        return named_outputs
-
-    def handleEvent(self, e):
-        """
-        Handles the arrival of a new event. Events are delivered from those
-        objects this DROP is subscribed to.
-        """
-        if e.type == "dropCompleted":
-            self.dropCompleted(e.uid, e.status)
-
-    def dropCompleted(self, uid, drop_state):
-        """
-        Callback invoked when the DROP with UID `uid` (which is either a
-        normal or a streaming input of this AppDROP) has moved to the
-        COMPLETED or ERROR state. By default no action is performed.
-        """
-
-    def dataWritten(self, uid, data):
-        """
-        Callback invoked when `data` has been written into the DROP with
-        UID `uid` (which is one of the streaming inputs of this AppDROP).
-        By default no action is performed
-        """
-
-    @property
-    def execStatus(self):
-        """
-        The execution status of this AppDROP
-        """
-        return self._execStatus
-
-    @execStatus.setter
-    def execStatus(self, execStatus):
-        if self._execStatus == execStatus:
-            return
-        self._execStatus = execStatus
-        self._fire("execStatus", execStatus=execStatus)
-
-    def _notifyAppIsFinished(self):
-        """
-        Method invoked by subclasses when the execution of the application is
-        over. Subclasses must make sure that both the status and execStatus
-        properties are set to their correct values correctly before invoking
-        this method.
-        """
-        is_error = self._execStatus == AppDROPStates.ERROR
-        if is_error:
-            self.status = DROPStates.ERROR
-        else:
-            self.status = DROPStates.COMPLETED
-        logger.debug(
-            "Moving %r to %s", self, "FINISHED" if not is_error else "ERROR"
-        )
-        self._fire(
-            "producerFinished", status=self.status, execStatus=self.execStatus
-        )
-        self.completedrop()
-
-    def cancel(self):
-        """Moves this application drop to its CANCELLED state"""
-        super(AppDROP, self).cancel()
-        self.execStatus = AppDROPStates.CANCELLED
-
-    def skip(self):
-        """Moves this application drop to its SKIPPED state"""
-        super().skip()
-
-        prev_execStatus = self.execStatus
-        self.execStatus = AppDROPStates.SKIPPED
-        for o in self._outputs.values():
-            o.skip()
-
-        logger.debug(f"Moving {self.__repr__()} to SKIPPED")
-        if prev_execStatus in [AppDROPStates.NOT_RUN]:
-            self._fire(
-                "producerFinished",
-                status=self.status,
-                execStatus=self.execStatus,
-            )
-
-
-class InputFiredAppDROP(AppDROP):
-    """
-    An InputFiredAppDROP accepts no streaming inputs and waits until a given
-    amount of inputs (called *effective inputs*) have moved to COMPLETED to
-    execute its 'run' method, which must be overwritten by subclasses. This way,
-    this application allows to continue the execution of the graph given a
-    minimum amount of inputs being ready. The transitions of subsequent inputs
-    to the COMPLETED state have no effect.
-
-    Normally only one call to the `run` method will happen per application.
-    However users can override this by specifying a different number of tries
-    before finally giving up.
-
-    The amount of effective inputs must be less or equal to the amount of inputs
-    added to this application once the graph is being executed. The special
-    value of -1 means that all inputs are considered as effective, in which case
-    this class acts as a BarrierAppDROP, effectively blocking until all its
-    inputs have moved to the COMPLETED, SKIPPED or ERROR state. Setting this
-    value to anything other than -1 or the number of inputs, results in
-    late arriving inputs to be ignored, even if they would successfully finish.
-    This requires careful implementation of the upstream and downstream apps to
-    deal with this situation. It is only really useful to control a combination
-    of maximum allowed execution time and acceptable number of completed inputs.
-
-    An input error threshold controls the behavior of the application given an
-    error in one or more of its inputs (i.e., a DROP moving to the ERROR state).
-    The threshold is a value within 0 and 100 that indicates the tolerance
-    to erroneous effective inputs, and after which the application will not be
-    run but moved to the ERROR state itself instead.
-    """
-
-    input_error_threshold = dlg_int_param(
-        "Input error threshold (0 and 100)", 0
-    )
-    n_effective_inputs = dlg_int_param("Number of effective inputs", -1)
-    n_tries = dlg_int_param("Number of tries", 1)
-
-    def initialize(self, **kwargs):
-        super(InputFiredAppDROP, self).initialize(**kwargs)
-        self._completedInputs = []
-        self._errorInputs = []
-        self._skippedInputs = []
-
-        # Error threshold must be within 0 and 100
-        if self.input_error_threshold < 0 or self.input_error_threshold > 100:
-            raise InvalidDropException(
-                self, "%r: input_error_threshold not within [0,100]" % (self,)
-            )
-
-        # Amount of effective inputs
-        if "n_effective_inputs" not in kwargs:
-            raise InvalidDropException(
-                self, "%r: n_effective_inputs is mandatory" % (self,)
-            )
-
-        if self.n_effective_inputs < -1 or self.n_effective_inputs == 0:
-            raise InvalidDropException(
-                self,
-                "%r: n_effective_inputs must be > 0 or equals to -1" % (self,),
-            )
-
-        # Number of tries
-        if self.n_tries < 1:
-            raise InvalidDropException(
-                self, "Invalid n_tries, must be a positive number"
-            )
-
-    def addStreamingInput(self, streamingInputDrop, back=True):
-        raise InvalidRelationshipException(
-            DROPRel(streamingInputDrop, DROPLinkType.STREAMING_INPUT, self),
-            "InputFiredAppDROPs don't accept streaming inputs",
-        )
-
-    def dropCompleted(self, uid, drop_state):
-        super(InputFiredAppDROP, self).dropCompleted(uid, drop_state)
-
-        logger.debug(
-            "Received notification from input drop: uid=%s, state=%d",
-            uid,
-            drop_state,
-        )
-
-        # A value of -1 means all inputs
-        n_inputs = len(self._inputs)
-        n_eff_inputs = self.n_effective_inputs
-        if n_eff_inputs == -1:
-            n_eff_inputs = n_inputs
-
-        # More effective inputs than inputs, this is a horror
-        if n_eff_inputs > n_inputs:
-            raise Exception(
-                "%r: More effective inputs (%d) than inputs (%d)"
-                % (self, self.n_effective_inputs, n_inputs)
-            )
-
-        if drop_state == DROPStates.ERROR:
-            self._errorInputs.append(uid)
-        elif drop_state == DROPStates.COMPLETED:
-            self._completedInputs.append(uid)
-        elif drop_state == DROPStates.SKIPPED:
-            self._skippedInputs.append(uid)
-        else:
-            raise Exception(
-                "Invalid DROP state in dropCompleted: %s" % drop_state
-            )
-
-        error_len = len(self._errorInputs)
-        ok_len = len(self._completedInputs)
-        skipped_len = len(self._skippedInputs)
-
-        # We have enough inputs to proceed
-        if (skipped_len + error_len + ok_len) == n_eff_inputs:
-
-            # calculate the number of errors that have already occurred
-            percent_failed = math.floor(
-                (error_len / float(n_eff_inputs)) * 100
-            )
-            if percent_failed > 0:
-                logger.debug(
-                    "Error rate on inputs for %r: %d/%d",
-                    self,
-                    percent_failed,
-                    self.input_error_threshold,
-                )
-
-            # if we hit the input error threshold then ERROR the drop and move on
-            if percent_failed > self.input_error_threshold:
-                logger.info(
-                    "Error threshold reached on %r, not executing it: %d/%d",
-                    self,
-                    percent_failed,
-                    self.input_error_threshold,
-                )
-
-                self.execStatus = AppDROPStates.ERROR
-                self.status = DROPStates.ERROR
-                self._notifyAppIsFinished()
-            elif skipped_len == n_eff_inputs:
-                self.skip()
-            else:
-                self.async_execute()
-
-    def async_execute(self):
-        # Return immediately, but schedule the execution of this app
-        # If we have been given a thread pool use that
-        if hasattr(self, "_tp"):
-            self._tp.apply_async(self._execute_and_log_exception)
-        else:
-            t = threading.Thread(target=self._execute_and_log_exception)
-            t.daemon = 1
-            t.start()
-            return t
-
-    def _execute_and_log_exception(self):
-        try:
-            self.execute()
-        except:
-            logger.exception(
-                "Unexpected exception during drop (%r) execution", self
-            )
-
-    _dlg_proc_lock = threading.Lock()
-
-    @track_current_drop
-    def execute(self, _send_notifications=True):
-        """
-        Manually trigger the execution of this application.
-
-        This method is normally invoked internally when the application detects
-        all its inputs are COMPLETED.
-        """
-
-        # TODO: We need to be defined more clearly how the state is set in
-        #       applications, for the time being they follow their execState.
-
-        # Run at most self._n_tries if there are errors during the execution
-        logger.debug("Executing %r", self)
-        tries = 0
-        drop_state = DROPStates.COMPLETED
-        self.execStatus = AppDROPStates.RUNNING
-        while tries < self.n_tries:
-            try:
-                if hasattr(self, "_tp"):
-                    proc = DlgProcess(target=self.run, daemon=True)
-                    # see YAN-975 for why this is happening
-                    lock = InputFiredAppDROP._dlg_proc_lock
-                    with lock:
-                        proc.start()
-                    with lock:
-                        proc.join()
-                    proc.close()
-                    if proc.exception:
-                        raise proc.exception
-                else:
-                    self.run()
-                if self.execStatus == AppDROPStates.CANCELLED:
-                    return
-                self.execStatus = AppDROPStates.FINISHED
-                break
-            except:
-                if self.execStatus == AppDROPStates.CANCELLED:
-                    return
-                tries += 1
-                logger.exception(
-                    "Error while executing %r (try %d/%d)",
-                    self,
-                    tries,
-                    self.n_tries,
-                )
-
-        # We gave up running the application, go to error
-        if tries == self.n_tries:
-            self.execStatus = AppDROPStates.ERROR
-            drop_state = DROPStates.ERROR
-
-        self.status = drop_state
-        if _send_notifications:
-            self._notifyAppIsFinished()
-
-    def run(self):
-        """
-        Run this application. It can be safely assumed that at this point all
-        the required inputs are COMPLETED.
-        """
-
-    # TODO: another thing we need to check
-    def exists(self):
-        return True
-
-
-class BarrierAppDROP(InputFiredAppDROP):
-    """
-    A BarrierAppDROP is an InputFireAppDROP that waits for all its inputs to
-    complete, effectively blocking the flow of the graph execution.
-    """
-
-    def initialize(self, **kwargs):
-        # Blindly override existing value if any
-        kwargs["n_effective_inputs"] = -1
-        super().initialize(**kwargs)
 
 
 # Dictionary mapping 1-to-many DROPLinkType constants to the corresponding methods
