+++ daliuge-engine/test/apps/test_bash.py
@@ -78,7 +78,9 @@ class BashAppTests(unittest.TestCase):
             a.addOutput(b)
             with DROPWaiterCtx(self, b, 100):
                 a.async_execute()
-            self.assertEqual(message.encode("utf8"), droputils.allDropContents(b))
+            self.assertEqual(
+                message.encode("utf8"), droputils.allDropContents(b)
+            )
 
         msg = "This is a message with a single quote: '"
         assert_message_is_correct(msg, 'echo -n "{0}" > %o0'.format(msg))
@@ -100,19 +102,23 @@ class BashAppTests(unittest.TestCase):
 
         def assert_envvar_is_there(varname, value):
             command = "echo -n $%s > %%o0" % (varname)
-            a = BashShellApp(app_uid, app_uid, dlg_session=session, command=command)
+            a = BashShellApp(
+                app_uid, app_uid, dlg_session=session, command=command
+            )
             b = FileDROP("b", "b")
             a.addOutput(b)
             with DROPWaiterCtx(self, b, 100):
                 a.async_execute()
-            self.assertEqual(value.encode("utf8"), droputils.allDropContents(b))
+            self.assertEqual(
+                value.encode("utf8"), droputils.allDropContents(b)
+            )
 
         assert_envvar_is_there("DLG_UID", app_uid)
         assert_envvar_is_there("DLG_SESSION_ID", session_id)
 
     def test_reproducibility(self):
         from dlg.common.reproducibility.constants import ReproducibilityFlags
-        from dlg.drop import NullDROP
+        from dlg.data.drops.data_base import NullDROP
 
         a = BashShellApp("a", "a", command="echo 'Hello world'")
         a.reproducibility_level = ReproducibilityFlags.RERUN
@@ -129,7 +135,9 @@ class BashAppTests(unittest.TestCase):
         a.reproducibility_level = ReproducibilityFlags.RECOMPUTE
         a.commit()
         self.assertNotEqual(a.merkleroot, b.merkleroot)
-        self.assertEqual(a.generate_merkle_data(), {"command": "echo 'Hello world'"})
+        self.assertEqual(
+            a.generate_merkle_data(), {"command": "echo 'Hello world'"}
+        )
 
         a.reproducibility_level = ReproducibilityFlags.REPRODUCE
         a.commit()
@@ -173,7 +181,9 @@ class StreamingBashAppTests(unittest.TestCase):
 
         output_fname = tempfile.mktemp()
 
-        a = StreamingOutputBashApp("a", "a", command=r"echo -en '5\n4\n3\n2\n1'")
+        a = StreamingOutputBashApp(
+            "a", "a", command=r"echo -en '5\n4\n3\n2\n1'"
+        )
         b = InMemoryDROP("b", "b")
         c = StreamingInputBashApp("c", "c", command="cat > %o0")
         d = FileDROP("d", "d", filepath=output_fname)
@@ -194,7 +204,8 @@ class StreamingBashAppTests(unittest.TestCase):
                 "Drop %r not COMPLETED: %d" % (drop, drop.status),
             )
         self.assertEqual(
-            [5, 4, 3, 2, 1], [int(x) for x in droputils.allDropContents(d).split(b"\n")]
+            [5, 4, 3, 2, 1],
+            [int(x) for x in droputils.allDropContents(d).split(b"\n")],
         )
 
         # Clean up and go
@@ -221,7 +232,9 @@ class StreamingBashAppTests(unittest.TestCase):
 
         output_fname = tempfile.mktemp()
 
-        a = StreamingOutputBashApp("a", "a", command=r"echo -en '5\n4\n3\n2\n1'")
+        a = StreamingOutputBashApp(
+            "a", "a", command=r"echo -en '5\n4\n3\n2\n1'"
+        )
         b = InMemoryDROP("b", "b")
         c = StreamingInputOutputBashApp("c", "c", command="cat")
         d = InMemoryDROP("d", "d")
@@ -243,7 +256,10 @@ class StreamingBashAppTests(unittest.TestCase):
             self.assertEqual(DROPStates.COMPLETED, drop.status)
         self.assertEqual(
             [1, 2, 3, 4, 5],
-            [int(x) for x in droputils.allDropContents(f).strip().split(b"\n")],
+            [
+                int(x)
+                for x in droputils.allDropContents(f).strip().split(b"\n")
+            ],
         )
 
         # Clean up and go
