+++ daliuge-engine/test/apps/test_dynlib.py
@@ -29,7 +29,7 @@ from dlg import droputils
 from dlg.apps.dynlib import DynlibApp, DynlibStreamApp, DynlibProcApp
 from dlg.common import Categories
 from dlg.ddap_protocol import DROPRel, DROPLinkType, DROPStates
-from dlg.drop import NullDROP
+from dlg.data.drops.data_base import NullDROP
 from dlg.data.drops.memory import InMemoryDROP
 
 from .setp_up import build_shared_library
@@ -43,7 +43,8 @@ bufsize = 20 * 1024 * 1024
 
 
 @unittest.skipUnless(
-    build_shared_library(_libname, _libpath), "Example dynamic library not available"
+    build_shared_library(_libname, _libpath),
+    "Example dynamic library not available",
 )
 class DynlibAppTest(unittest.TestCase):
     def test_simple_batch_copy(self):
@@ -87,7 +88,9 @@ class DynlibAppTest(unittest.TestCase):
             if streaming:
                 # Write the data in chunks so we actually exercise multiple calls
                 # to the data_written library call
-                for datum in iter(functools.partial(reader.read, 1024 * 1024), b""):
+                for datum in iter(
+                    functools.partial(reader.read, 1024 * 1024), b""
+                ):
                     a.write(datum)
             else:
                 a.write(data)
@@ -114,7 +117,9 @@ class DynlibAppTest(unittest.TestCase):
         t0 = time.time()
         a.cancel()
         self.assertLess(
-            time.time() - t0, 1, "Cancelled dynlibprocapp in less than a second"
+            time.time() - t0,
+            1,
+            "Cancelled dynlibprocapp in less than a second",
         )
         self.assertEqual(DROPStates.CANCELLED, a.status)
 
@@ -223,19 +228,27 @@ class IntraNMMixIng(test_dm.NMTestsMixIn):
         ]
         a_data = os.urandom(32)
         self._test_runGraphInTwoNMs(
-            g1, g2, rels, a_data, a_data * 2, root_oids=("A", "B"), leaf_oid="D"
+            g1,
+            g2,
+            rels,
+            a_data,
+            a_data * 2,
+            root_oids=("A", "B"),
+            leaf_oid="D",
         )
 
 
 @unittest.skipUnless(
-    build_shared_library(_libname, _libpath), "Example dynamic library not available"
+    build_shared_library(_libname, _libpath),
+    "Example dynamic library not available",
 )
 class IntraNMDynlibAppTest(IntraNMMixIng, unittest.TestCase):
     app = "dlg.apps.dynlib.DynlibApp"
 
 
 @unittest.skipUnless(
-    build_shared_library(_libname, _libpath), "Example dynamic library not available"
+    build_shared_library(_libname, _libpath),
+    "Example dynamic library not available",
 )
 class IntraNMDynlibProcAppTest(IntraNMMixIng, unittest.TestCase):
     app = "dlg.apps.dynlib.DynlibProcApp"
