+++ daliuge-engine/test/lifecycle/test_dlm.py
@@ -32,7 +32,7 @@ import time
 import unittest
 
 from dlg.ddap_protocol import DROPStates, DROPPhases
-from dlg.drop import BarrierAppDROP
+from dlg.apps.app_base import BarrierAppDROP
 from dlg.data.drops.directorycontainer import DirectoryContainer
 from dlg.data.drops.file import FileDROP
 from dlg.droputils import DROPWaiterCtx
@@ -108,7 +108,9 @@ class TestDataLifecycleManager(unittest.TestCase):
             self.assertEqual(DROPPhases.LOST, drop.phase)
 
     def test_cleanupExpiredDrops(self):
-        with dlm.DataLifecycleManager(check_period=0.5, cleanup_period=2) as manager:
+        with dlm.DataLifecycleManager(
+            check_period=0.5, cleanup_period=2
+        ) as manager:
             drop = FileDROP(
                 "oid:A", "uid:A1", expectedSize=1, lifespan=1, persist=False
             )
@@ -136,7 +138,9 @@ class TestDataLifecycleManager(unittest.TestCase):
         different values, and after they are used we check whether their data
         is still there or not
         """
-        with dlm.DataLifecycleManager(check_period=0.5, cleanup_period=2) as manager:
+        with dlm.DataLifecycleManager(
+            check_period=0.5, cleanup_period=2
+        ) as manager:
             a = DirectoryContainer(
                 "a",
                 "a",
@@ -146,7 +150,11 @@ class TestDataLifecycleManager(unittest.TestCase):
             )
             b_dirname = tempfile.mkdtemp()
             b = DirectoryContainer(
-                "b", "b", persist=False, expireAfterUse=False, dirname=b_dirname
+                "b",
+                "b",
+                persist=False,
+                expireAfterUse=False,
+                dirname=b_dirname,
             )
             c = BarrierAppDROP("c", "c")
             d = BarrierAppDROP("d", "d")
