+++ daliuge-engine/test/manager/test_daemon.py
@@ -30,11 +30,13 @@ from dlg.manager import constants
 from dlg.manager.client import MasterManagerClient
 from dlg.manager.proc_daemon import DlgDaemon
 
-_TIMEOUT = 30
+_TIMEOUT = 5
 IDENTITY = lambda x: x
 
 
-def wait_until(update_condition, test_condition=IDENTITY, timeout=_TIMEOUT, interval=0.1):
+def wait_until(
+    update_condition, test_condition=IDENTITY, timeout=_TIMEOUT, interval=0.1
+):
     timeout_time = time.time() + timeout
     while time.time() < timeout_time:
         output = update_condition()
@@ -92,7 +94,9 @@ class TestDaemon(unittest.TestCase):
         # in the daemon's hand
         # self.assertTrue(utils.portIsOpen('localhost', 9000, _TIMEOUT))
         try:
-            restutils.RestClient("localhost", 9000, timeout=10)._GET("/anything")
+            restutils.RestClient("localhost", 9000, timeout=_TIMEOUT)._GET(
+                "/anything"
+            )
         except restutils.RestClientException:
             # We don't care about the result
             pass
@@ -123,11 +127,15 @@ class TestDaemon(unittest.TestCase):
         # Nothing should start now
         self.create_daemon(master=False, noNM=True, disable_zeroconf=True)
         self.assertTrue(
-            utils.portIsClosed("localhost", constants.NODE_DEFAULT_REST_PORT, 0),
+            utils.portIsClosed(
+                "localhost", constants.NODE_DEFAULT_REST_PORT, 0
+            ),
             "NM started but it should not have",
         )
         self.assertTrue(
-            utils.portIsClosed("localhost", constants.MASTER_DEFAULT_REST_PORT, 0),
+            utils.portIsClosed(
+                "localhost", constants.MASTER_DEFAULT_REST_PORT, 0
+            ),
             "NM started but it should not have",
         )
 
@@ -149,13 +157,16 @@ class TestDaemon(unittest.TestCase):
 
     def _test_zeroconf_dim_mm(self, disable_zeroconf=False):
         # Start an empty daemon, then a DIM and a Master on their own
-        self.create_daemon(master=False, noNM=True, disable_zeroconf=disable_zeroconf)
+        self.create_daemon(
+            master=False, noNM=True, disable_zeroconf=disable_zeroconf
+        )
         self._start("island", http.HTTPStatus.OK, {"nodes": []})
         self._start("master", http.HTTPStatus.OK)
 
         # Check that dim registers to MM when using zeroconf
         mc = MasterManagerClient()
         if not disable_zeroconf:
+
             def _test_dims(dims):
                 return dims and dims["islands"]
 
@@ -197,7 +208,7 @@ class TestDaemon(unittest.TestCase):
 
     def test_without_zeroconf_dim_nm_setup(self):
         self._test_zeroconf_dim_mm(disable_zeroconf=True)
-        nodes = self._add_zeroconf_nm()['nodes']
+        nodes = self._add_zeroconf_nm()["nodes"]
         self.assertEqual(0, len(nodes))
 
     def test_zeroconf_nm_down(self):
@@ -208,9 +219,11 @@ class TestDaemon(unittest.TestCase):
         mc = MasterManagerClient()
 
         def _test_nodes(nodes):
-            return not nodes['nodes']
+            return not nodes["nodes"]
 
-        new_nodes = _get_nodes_from_client(mc, test_condition=_test_nodes)['nodes']
+        new_nodes = _get_nodes_from_client(mc, test_condition=_test_nodes)[
+            "nodes"
+        ]
         self.assertEqual(0, len(new_nodes))
 
     def test_start_dataisland_via_rest(self):
@@ -232,7 +245,9 @@ class TestDaemon(unittest.TestCase):
         # Check that the DataIsland starts with the given nodes
         self._start("island", http.HTTPStatus.OK, {"nodes": nodes})
         self.assertTrue(
-            utils.portIsOpen("localhost", constants.ISLAND_DEFAULT_REST_PORT, _TIMEOUT),
+            utils.portIsOpen(
+                "localhost", constants.ISLAND_DEFAULT_REST_PORT, _TIMEOUT
+            ),
             "The DIM did not start successfully",
         )
 
@@ -259,7 +274,9 @@ class TestDaemon(unittest.TestCase):
         # Check that the DataIsland stopped
         self._stop("island", http.HTTPStatus.OK, "")
         self.assertTrue(
-            utils.portIsClosed("localhost", constants.ISLAND_DEFAULT_REST_PORT, _TIMEOUT),
+            utils.portIsClosed(
+                "localhost", constants.ISLAND_DEFAULT_REST_PORT, _TIMEOUT
+            ),
             "The DIM did not stop successfully",
         )
 
@@ -283,14 +300,18 @@ class TestDaemon(unittest.TestCase):
         # Check that the NM stops
         self._stop("node", http.HTTPStatus.OK, "")
         self.assertTrue(
-            utils.portIsClosed('localhost', constants.NODE_DEFAULT_REST_PORT, _TIMEOUT),
+            utils.portIsClosed(
+                "localhost", constants.NODE_DEFAULT_REST_PORT, _TIMEOUT
+            ),
             "The node did not stop successfully",
         )
 
         # Check that the NM starts
         self._start("node", http.HTTPStatus.OK, {"pid": nodes})
         self.assertTrue(
-            utils.portIsOpen('localhost', constants.NODE_DEFAULT_REST_PORT, _TIMEOUT),
+            utils.portIsOpen(
+                "localhost", constants.NODE_DEFAULT_REST_PORT, _TIMEOUT
+            ),
             "The node did not start successfully",
         )
 
@@ -301,14 +322,18 @@ class TestDaemon(unittest.TestCase):
         # Check that the MM starts
         self._start("master", http.HTTPStatus.OK)
         self.assertTrue(
-            utils.portIsOpen("localhost", constants.MASTER_DEFAULT_REST_PORT, _TIMEOUT),
+            utils.portIsOpen(
+                "localhost", constants.MASTER_DEFAULT_REST_PORT, _TIMEOUT
+            ),
             "The MM did not start successfully",
         )
 
         # Check that the MM stops
         self._stop("master", http.HTTPStatus.OK, "")
         self.assertTrue(
-            utils.portIsClosed("localhost", constants.MASTER_DEFAULT_REST_PORT, _TIMEOUT),
+            utils.portIsClosed(
+                "localhost", constants.MASTER_DEFAULT_REST_PORT, _TIMEOUT
+            ),
             "The MM did not stop successfully",
         )
 
@@ -348,7 +373,10 @@ class TestDaemon(unittest.TestCase):
             payload = json.dumps(payload)
             headers["Content-Type"] = "application/json"
         conn.request(
-            "POST", f"/managers/{manager_name}/stop", body=payload, headers=headers
+            "POST",
+            f"/managers/{manager_name}/stop",
+            body=payload,
+            headers=headers,
         )
         response = conn.getresponse()
         self.assertEqual(expected_code, response.status, response.read())
